# servers/fhir-clinical/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "FHIR Clinical (curated MCP subset)",
    "version": "1.0.0",
    "description": "A curated clinical subset of HAPI FHIR R4 exposed as MCP tools. Read/search only. The full HAPI OpenAPI is ~3,267 operations; this subset is what the FHIR clinical MCP server fronts."
  },
  "servers": [{ "url": "http://hapi:8081/fhir" }],
  "paths": {
    "/Patient/{id}": {
      "get": {
        "operationId": "getPatient",
        "summary": "Read a Patient by id",
        "description": "Read a single Patient resource. Patient-scoped callers may only read their own record.",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" }, "description": "FHIR Patient logical id" }
        ],
        "responses": { "200": { "description": "Patient resource" } }
      }
    },
    "/Patient/{id}/$everything": {
      "get": {
        "operationId": "patientEverything",
        "summary": "Everything in a patient's compartment",
        "description": "Return the full patient compartment. Broad operation: clinical role + step-up scope required.",
        "x-mcp-group": "mcp-clinical-tools",
        "x-mcp-scope": "mcp:fhir-clinical:everything:read",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" }, "description": "FHIR Patient logical id" }
        ],
        "responses": { "200": { "description": "Bundle" } }
      }
    },
    "/Observation": {
      "get": {
        "operationId": "searchObservation",
        "summary": "Search Observations",
        "parameters": [
          { "name": "patient", "in": "query", "schema": { "type": "string" }, "description": "Patient reference (id)" },
          { "name": "code", "in": "query", "schema": { "type": "string" }, "description": "Observation code (LOINC)" },
          { "name": "_count", "in": "query", "schema": { "type": "integer" }, "description": "Max results" }
        ],
        "responses": { "200": { "description": "Bundle" } }
      }
    },
    "/Condition": {
      "get": {
        "operationId": "searchCondition",
        "summary": "Search Conditions",
        "parameters": [
          { "name": "patient", "in": "query", "schema": { "type": "string" }, "description": "Patient reference (id)" },
          { "name": "_count", "in": "query", "schema": { "type": "integer" }, "description": "Max results" }
        ],
        "responses": { "200": { "description": "Bundle" } }
      }
    },
    "/MedicationRequest": {
      "get": {
        "operationId": "searchMedicationRequest",
        "summary": "Search MedicationRequests",
        "parameters": [
          { "name": "patient", "in": "query", "schema": { "type": "string" }, "description": "Patient reference (id)" },
          { "name": "_count", "in": "query", "schema": { "type": "integer" }, "description": "Max results" }
        ],
        "responses": { "200": { "description": "Bundle" } }
      }
    }
  }
}